# app/schemas/responses.py
"""
Response schemas.
"""
from pydantic import BaseModel, Field
from typing import List, Dict, Literal
from app.models.domain import DomainType

class Variation(BaseModel):
    id: int
    question: str
    answer_key: str
    difficulty: str
    question_type: str
    topic: str
    subtopic: str
    confidence_score: float = Field(..., ge=0, le=1)
    flagged_for_review: bool

class GenerateResponse(BaseModel):
    seed_question: str
    domain: DomainType
    total_generated: int
    duplicate_rate: float
    variations: List[Variation]
    review_queue: List[Variation]
    job_id: str

class DomainInfo(BaseModel):
    name: str
    label: str
    description: str
    supported_question_types: List[str]
    topics: Dict[str, List[str]]

class DomainsResponse(BaseModel):
    domains: List[DomainInfo]

class HealthResponse(BaseModel):
    status: Literal['healthy', 'degraded', 'unhealthy']
    ollama_connected: bool
    model_loaded: str
    embedding_model_loaded: bool
    version: str

class ExportResponse(BaseModel):
    job_id: str
    format: Literal['csv', 'json']
    content: str
    filename: str