# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Poly Prompt Engine",
    "description": "Assignment Question Iteration & Variation Generation System. Generates domain-aware question variations from a seed question using local open-weight LLMs. National Hackathon 2026 \u2014 PS-8.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/export/{job_id}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Job",
        "description": "Export generated variations for a specific job.",
        "operationId": "export_job_api_v1_export__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Export format: json or csv",
              "default": "json",
              "title": "Format"
            },
            "description": "Export format: json or csv"
          },
          {
            "name": "download",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Return as a file download",
              "default": false,
              "title": "Download"
            },
            "description": "Return as a file download"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ExportResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "format": {
            "type": "string",
            "enum": [
              "csv",
              "json"
            ],
            "title": "Format"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "format",
          "content",
          "filename"
        ],
        "title": "ExportResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}